Refactor push
Added exception handling to the font lookups.

Changed screen size (still hard-coded). Edited import lines.

Added variables to reduce the amount of hard-coding, and to work towards the S and O principles of SOLID.

As segments and apples are square, we only need one dimension for them instead of 2. Thus, I have changed the dimension lists to variables.